moved some helper file functions
